Stopped at weird template compilation error.

+++ caffe2/contrib/aten/aten_op.cc
@@ -18,7 +18,7 @@ void Set<at::Half, CPUContext>(
     const at::Half h,
     at::Half* v,
     CPUContext* c) {
-  Set(0, h.x, (uint16_t*) v, c);
+  Set(0, h.x, (uint32_t*) v, c);
 }
 }
